requirements.txt
fastapi
pydantic
pydantic[email]
python-dotenv
pydantic-settings
resend
PyJWT
bcrypt
redis
uvicorn
psycopg2-binary


langchain
pinecone-client
langchain-pinecone
langchain-openai
pypdf
transformers